Move document spec into DocumentTypes directly, and split project and traj version
